+++ argilla-server/src/argilla_server/api/schemas/v1/workspaces.py
@@ -13,7 +13,8 @@
 #  limitations under the License.
 
 from datetime import datetime
-from typing import List
+from optparse import Option
+from typing import List, Optional
 from uuid import UUID
 
 from pydantic import BaseModel, Field, ConfigDict
@@ -29,6 +30,7 @@ class Workspace(BaseModel):
 
 
 class WorkspaceCreate(BaseModel):
+    id: Optional[UUID] = None
     name: str = Field(min_length=1)