+++ argilla-server/src/argilla_server/validators/users.py
@@ -22,6 +22,7 @@ class UserCreateValidator:
     @classmethod
     async def validate(cls, db: AsyncSession, user: User) -> None:
         await cls._validate_username(db, user)
+        await cls._validate_user_id(db, user)
 
     @classmethod
     async def _validate_username(cls, db, user: User):
@@ -39,3 +40,11 @@ class UserCreateValidator:
     async def _validate_username_length(cls, user: User):
         if len(user.username) < 1:
             raise UnprocessableEntityError("Username must be at least 1 characters long")
+
+    @classmethod
+    async def _validate_user_id(cls, db, user: User):
+        if user.id is None:
+            return
+
+        if await User.get(db, id=user.id) is not None:
+            raise NotUniqueError(f"User with id `{user.id}` is not unique")