+++ argilla-server/tests/unit/api/handlers/v1/test_datasets.py
@@ -1479,7 +1479,7 @@ class TestSuiteDatasets:
         assert (await db.execute(select(func.count(Response.id)))).scalar() == 4
         assert (await db.execute(select(func.count(Suggestion.id)))).scalar() == 3
 
-        records = (await db.execute(select(Record))).scalars().all()
+        records = (await db.execute(select(Record).order_by(Record.inserted_at.asc()))).scalars().all()
         mock_search_engine.index_records.assert_called_once_with(dataset, records)
 
     async def test_create_dataset_records_with_response_for_multiple_users(
@@ -1547,7 +1547,7 @@ class TestSuiteDatasets:
         assert owner in dataset.users
         assert annotator in dataset.users
 
-        records = (await db.execute(select(Record))).scalars().all()
+        records = (await db.execute(select(Record).order_by(Record.inserted_at.asc()))).scalars().all()
         mock_search_engine.index_records.assert_called_once_with(dataset, records)
 
     async def test_create_dataset_records_with_response_for_unknown_user(
@@ -2109,7 +2109,9 @@ class TestSuiteDatasets:
         assert (await db.execute(select(func.count(Vector.id)))).scalar() == 3
 
         vector_a, vector_b, vector_c = (await db.execute(select(Vector))).scalars().all()
-        record_a, record_b, record_c = (await db.execute(select(Record))).scalars().all()
+        record_a, record_b, record_c = (
+            (await db.execute(select(Record).order_by(Record.inserted_at.asc()))).scalars().all()
+        )
         assert (
             vector_a.record_id == record_a.id
             and vector_a.vector_settings_id == vector_settings_a.id
@@ -2314,7 +2316,7 @@ class TestSuiteDatasets:
         assert (await db.execute(select(func.count(Record.id)))).scalar() == 5
         assert (await db.execute(select(func.count(Response.id)))).scalar() == 4
 
-        records = (await db.execute(select(Record))).scalars().all()
+        records = (await db.execute(select(Record).order_by(Record.inserted_at.asc()))).scalars().all()
         mock_search_engine.index_records.assert_called_once_with(dataset, records)
 
     async def test_create_dataset_records_as_annotator(self, async_client: "AsyncClient", db: "AsyncSession"):