+++ argilla/src/argilla/users/_resource.py
@@ -45,8 +45,8 @@ class User(Resource):
         last_name: Optional[str] = None,
         role: Optional[str] = None,
         password: Optional[str] = None,
-        client: Optional["Argilla"] = None,
         id: Optional[UUID] = None,
+        client: Optional["Argilla"] = None,
         _model: Optional[UserModel] = None,
     ) -> None:
         """Initializes a User object with a client and a username
@@ -57,6 +57,7 @@ class User(Resource):
             last_name (str): The last name of the user
             role (str): The role of the user, either 'annotator', admin, or 'owner'
             password (str): The password of the user
+            id (UUID): The ID of the user. If provided before a .create, the will be created with this ID
             client (Argilla): The client used to interact with Argilla
 
         Returns: